Add first version of neural net experiment, make all dataset features float32

+++ src/concrete_ml_playground/datasets/sms_spam.py
@@ -11,4 +11,4 @@ def load_sms_spam_dataset(max_features: None | int = None) -> tuple[npt.NDArray,
     vectorizer = TfidfVectorizer(stop_words="english", max_features=max_features)
     X = vectorizer.fit_transform(dataset["sms"])
 
-    return X.toarray(), np.array(dataset["label"])
+    return X.toarray().astype(np.float32), np.array(dataset["label"])
