+++ src/concrete_ml_playground/datasets/xor.py
@@ -2,7 +2,7 @@ import numpy as np
 
 
 def load_xor_split_dataset():
-    X = np.array([[0.25, 0.25], [0.25, 0.75], [0.75, 0.25], [0.75, 0.75]])
+    X = np.array([[0.25, 0.25], [0.25, 0.75], [0.75, 0.25], [0.75, 0.75]], dtype=np.float32)
     y = np.array([0, 1, 1, 0])
 
     # some models expect datasets of a certain size, so we just put the same samples in multiple times
@@ -15,8 +15,8 @@ def load_xor_split_dataset():
     X_test, y_test = X[p2], y[p2]
 
     # add some noise so that train and test set are not identical and to make them floats
-    noise1 = rng.uniform(-0.24, 0.24, (len(X), 2))
-    noise2 = rng.uniform(-0.24, 0.24, (len(X), 2))
+    noise1 = rng.uniform(-0.24, 0.24, (len(X), 2)).astype(np.float32)
+    noise2 = rng.uniform(-0.24, 0.24, (len(X), 2)).astype(np.float32)
     X_train = np.add(X_train, noise1)
     X_test = np.add(X_test, noise2)
